refactor(e2e): own Redis E2E in driver package; drop Host Kiwi spec
Default wdio specs stay under e2e/specs; e2e:redis is explicit-only.
