+++ package.json
@@ -22,8 +22,8 @@
     "e2e:wdio": "wdio run e2e/wdio.conf.ts",
     "e2e:core": "node e2e/run.mjs --skip-build -- --spec e2e/specs/main-window.ts,e2e/specs/new-connection.ts,e2e/specs/edit-delete-connection.ts,e2e/specs/connection-search-group.ts,e2e/specs/settings.ts,e2e/specs/i18n-menu.ts,e2e/specs/homepage-features.ts,e2e/specs/drag-drop-groups.ts,e2e/specs/backup-database.ts",
     "e2e:db": "node e2e/run.mjs --skip-build -- --spec e2e/specs/connection-window.ts,e2e/specs/sql-query.ts,e2e/specs/table-data.ts,e2e/specs/table-edit.ts,e2e/specs/table-structure.ts,e2e/specs/export-import.ts,e2e/specs/data-types.ts,e2e/specs/mysql.ts,e2e/specs/mysql-multi-database.ts,e2e/specs/postgres-multi-database.ts,e2e/specs/data-sync-real.ts",
-    "e2e:kiwi": "node e2e/run.mjs --skip-build -- --spec e2e/specs/kiwi.ts",
-    "e2e:redis": "node e2e/run.mjs --skip-build -- --spec e2e/specs/redis.ts,e2e/specs/redis-topology.ts",
+    "e2e:kiwi": "node -e \"console.error('Kiwi E2E lives in the datazen-driver-kiwi repo; Host no longer ships e2e/specs/kiwi.ts'); process.exit(1)\"",
+    "e2e:redis": "node e2e/run.mjs --skip-build -- --spec packages/drivers/redis/e2e/redis.ts,packages/drivers/redis/e2e/redis-topology.ts",
     "e2e:ai": "node e2e/run.mjs --skip-build -- --spec e2e/specs/ai-features.ts,e2e/specs/ai-context.ts,e2e/specs/ai-context-tables.ts",
     "e2e:i18n-backup": "node e2e/run.mjs --skip-build -- --spec e2e/specs/app-data-backup.ts,e2e/specs/i18n-10-locales.ts,e2e/specs/system-locale.ts,e2e/specs/i18n-menu.ts",
     "e2e:path-ipc": "node e2e/run.mjs --skip-build -- --spec e2e/specs/path-ipc-hardening.ts,e2e/specs/workflow-window.ts,e2e/specs/app-data-backup.ts",
